+++ shardis-ui/src/main/frontend/app/about/about.component.spec.ts
@@ -1,7 +1,6 @@
 /* tslint:disable:no-unused-variable */
 
-import { TestBed, async } from '@angular/core/testing';
-import { AboutComponent } from './about.component';
+import {AboutComponent} from './about.component';
 
 describe('Component: About', () => {
   it('should create an instance', () => {