+++ shardis-ui/src/main/frontend/app/app.module.ts
@@ -26,19 +26,17 @@ import {
 import {HomeComponent} from './home/home.component';
 import {NotFoundComponent} from './not-found/not-found.component';
 import {AccessDeniedComponent} from './access-denied/access-denied.component';
-import {LoginComponent} from './login/login.component';
 import {AuthService} from './shared/auth/auth.service';
 import {AdminGuard} from './shared/guards/admin.guard';
 import {AuthenticatedGuard} from './shared/guards/authenticated.guard';
 import {UnauthenticatedGuard} from './shared/guards/unauthenticated.guard';
 import {TitleService} from './home/shared/title.service';
-import {ShardisUiRoutingModule} from './app-routing.module';
+import {AppRoutingModule} from './app-routing.module';
 
 @NgModule({
   declarations: [
     AppComponent,
     HomeComponent,
-    LoginComponent,
     NotFoundComponent,
     AccessDeniedComponent
   ],
@@ -64,7 +62,7 @@ import {ShardisUiRoutingModule} from './app-routing.module';
     MdTabsModule.forRoot(),
     MdToolbarModule.forRoot(),
     MdTooltipModule.forRoot(),
-    ShardisUiRoutingModule
+    AppRoutingModule
   ],
   providers: [
     AuthService,