+++ shardis-ui/src/main/frontend/app/not-found/not-found.component.spec.ts
@@ -1,7 +1,6 @@
 /* tslint:disable:no-unused-variable */
 
-import { TestBed, async } from '@angular/core/testing';
-import { NotFoundComponent } from './not-found.component';
+import {NotFoundComponent} from './not-found.component';
 
 describe('Component: NotFound', () => {
   it('should create an instance', () => {