+++ shardis-ui/src/main/frontend/app/playground/playground.component.spec.ts
@@ -1,7 +1,6 @@
 /* tslint:disable:no-unused-variable */
 
-import { TestBed, async } from '@angular/core/testing';
-import { PlaygroundComponent } from './playground.component';
+import {PlaygroundComponent} from './playground.component';
 
 describe('Component: Playground', () => {
   it('should create an instance', () => {