+++ shardis-ui/src/main/frontend/app/playground/shared/hero-detail/hero-detail.component.spec.ts
@@ -1,7 +1,6 @@
 /* tslint:disable:no-unused-variable */
 
-import { TestBed, async } from '@angular/core/testing';
-import { HeroDetailComponent } from './hero-detail.component';
+import {HeroDetailComponent} from './hero-detail.component';
 
 describe('Component: HeroDetail', () => {
   it('should create an instance', () => {