+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@
 * Updated Docker Maven Plugin to 0.4.13
 * Added docker cleanup script for Windows
 * Added tags for generated docker images
+* Added separate routing modules to all lazy loaded Angular modules
 
 ### 2.1.0 (28.08.2016)
 * Added Spring Cloud Config server for centralized configuration